add sampling approach solution

+++ SamplingApproach.cpp
@@ -6,7 +6,7 @@
 
 std::vector<int> SamplingApproach::solve(Grapher &graph)  {
     std::vector<int> x(graph.getM(), 0);
-    std::vector<std::pair<Path, long double> > P;
+    std::vector<std::pair<Path, double> > P;
     while (true) {
         P = graph.samplingPath(x);
         if (P.empty()) {
@@ -17,7 +17,7 @@ std::vector<int> SamplingApproach::solve(Grapher &graph)  {
     return x;
 }
 
-void SamplingApproach::greedyAlgorithm(Grapher &graph, std::vector<std::pair<Path, long double> > &P, std::vector<int> &x, int q) {
+void SamplingApproach::greedyAlgorithm(Grapher &graph, std::vector<std::pair<Path, double> > &P, std::vector<int> &x, int q) {
     std::vector<bool> checked(graph.getM(), false);
     for (int i = 0; i < P.size(); i++) {
         for (int j : P[i].first) {
@@ -26,12 +26,12 @@ void SamplingApproach::greedyAlgorithm(Grapher &graph, std::vector<std::pair<Pat
     }
     while (q--) {
         int pos = -1;
-        long double res = 0;
+        double res = 0;
         for (int i = graph.getM() - 1; i >= 0; i--) {
             if (!checked[i] || x[i] == graph.getEdge(i).getWeight().size() - 1) {
                 continue;
             }
-            long double pot = 0;
+            double pot = 0;
             std::vector<int> newx = addition(x, i, 1);
             for (int j = 0; j < P.size(); j++) {
                 int tmp = graph.budgetFunction(P[j].first, newx);
