+++ SamplingApproach.h
@@ -8,7 +8,7 @@
 
 
 class SamplingApproach {
-    static void greedyAlgorithm(Grapher &graph, std::vector<std::pair<Path, long double> > &P, std::vector<int> &x, int q);
+    static void greedyAlgorithm(Grapher &graph, std::vector<std::pair<Path, double> > &P, std::vector<int> &x, int q);
 public:
     static std::vector<int> solve(Grapher &graph);
 };
